GS-9825 Changed cursor from int to long
svn path=/xap/trunk/openspaces/; revision=106608


Former-commit-id: 08d4eb6a688819d31a97f0c204d56b692e4bc778

+++ src/main/src/org/openspaces/admin/internal/gsm/DefaultGridServiceManager.java
@@ -78,7 +78,7 @@ public class DefaultGridServiceManager extends AbstractAgentGridComponent implem
 
     private final ProvisionMonitorAdmin gsmAdmin;
 
-    private int eventsCursor = 0;
+    private long eventsCursor = 0;
 
     public DefaultGridServiceManager(ServiceID serviceID, GSM gsm, InternalAdmin admin, int agentId, String agentUid)
             throws RemoteException {
